Fixing exit overflow leak 215

+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jkoupy <jkoupy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jseidere <jseidere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/09 12:04:06 by jkoupy            #+#    #+#             */
-/*   Updated: 2024/05/15 12:05:00 by jkoupy           ###   ########.fr       */
+/*   Updated: 2024/05/15 14:36:39 by jseidere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -143,7 +143,7 @@ void	exit_command(t_shell *shell, t_cmd *cmd);
 //exit_util.c
 bool	check_amount_of_args(char **args);
 bool	is_numeric(char *str);
-char	*ft_ltoa(long n);
+char	*ft_ltoa(long long n);
 long	convert_exit_status(t_cmd *cmd);
 bool	check_overflow(char *str);
