+++ library/libft/include/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jkoupy <jkoupy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jseidere <jseidere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 10:54:04 by jkoupy            #+#    #+#             */
-/*   Updated: 2024/04/22 12:37:20 by jkoupy           ###   ########.fr       */
+/*   Updated: 2024/05/15 14:36:28 by jseidere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,7 +93,7 @@ char	*ft_dtoa(double n, int precision);
 char	*ft_strjoin3(char const *s1, char const *s2, char const *s3);
 char	*ft_strlwr(char *str);
 long	ft_atol(const char *nptr);
-char	*ft_ltoa(long n);
+char	*ft_ltoa(long long n);
 int		ft_isspace(int c);
 
 #endif
